added exercises 2 b through e
